get cloud kinda-sync kinda working
unstarred files are essentially temp files. They are only persisted locally. They are auto-saved (locally) as you edit them. When the user "stars" a mixture, they will cause that local file to be pushed to R2. As the user edits a starred mixture, we autosave locally, and we show a "dirty" indicator in the UI when the local mixture hash diverges from the remote hash. When the user clicks to save the mixture, we push it to R2 (and here if the version in R2 is newer than the last time we saved the mixture, we show a confirmation prompt before overwriting the R2 object).