refactor collections initialization & separate mixture-list from ingredient list collections
The way signaldb works, the entire dataset is transmitted twice whenever any item in the set changes. This meant that in the old scheme, the user's entire library was transmitted 2x each time they edited an ingredient in an individual mixture. This change splits off the mixture metadata (which shouldn't change as frequently) into one collection, and it creates individual collections for each mixture in the user's library. That way, when the user edits an ingredient in a mixture, only that mixture's ingredient data needs to be transmitted 2x on each change.